requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.1
Werkzeug==2.0.1
Jinja2==3.1.3
Markdown==3.5.2
gunicorn==20.1.0 